Added a union of four function pointers, their setter methods and applyPendingChanges() function; amended changeState() function
Function pointers will hold an address copy of a function to call to perform a safe game state transition before the start of the next game state update cycle.
applyPendingChanges() function will choose which function pointer to use based on the current game state.
changeState() will now completely clear the state stack before pushing a new one to avoid a scenario when a new PlayState is added on top of previous PlayState after GameOverState was removed